+++ src/compas_model/modifiers/__init__.py
@@ -1,7 +1,5 @@
 from .modifier import Modifier
-from .trim import TrimModifier
 
 __all__ = [
     "Modifier",
-    "TrimModifier",
 ]